renamed kuzzleSecurity.js -> Security.js
